+++ rendering/renderer.go
@@ -2,9 +2,9 @@ package rendering
 
 type Renderer interface {
 	Prepare()
-	AddPrimitive() int64
-	SetPrimitive(id int64, primitive IPrimitive, shouldRerender bool)
-	RemovePrimitive(id int64)
+	AddPrimitive() int
+	SetPrimitive(id int, primitive IPrimitive, shouldRerender bool)
+	RemovePrimitive(id int)
 	PerformRendering()
 	Clear()
 	Stop()
